Added to export all theme.

+++ vite.config.mts
@@ -1,6 +1,6 @@
 import { resolve } from 'node:path';
 import { defineConfig } from 'vite';
-import { copyFileSync, unlinkSync } from 'node:fs';
+import { copyFileSync, unlinkSync, existsSync } from 'node:fs';
 import { fileURLToPath, URL } from 'url';
 import vue from '@vitejs/plugin-vue';
 import dts from 'vite-plugin-dts';
@@ -16,6 +16,11 @@ const config = {
         entry: resolve(__dirname, './src/styles.ts'),
         fileName: 'styles.js',
     },
+    theme: {
+        name: 'theme',
+        entry: resolve(__dirname, './src/theme.ts'),
+        fileName: 'theme.js',
+    },
 };
 
 const currentConfig = config[process.env.LIB_NAME];
@@ -24,23 +29,23 @@ if (currentConfig === undefined) {
     throw new Error('LIB_NAME is not defined or is not valid');
 }
 
+const removeUnnecessary = () => {
+    if (existsSync('dist/styles.d.ts')) copyFileSync('dist/index.d.ts', 'dist/index.d.mts');
+    if (existsSync('dist/styles.d.ts')) unlinkSync('dist/styles.d.ts');
+    if (existsSync('dist/styles.js.mjs')) unlinkSync('dist/styles.js.mjs');
+    if (existsSync('dist/styles.js.umd.js')) unlinkSync('dist/styles.js.umd.js');
+    if (existsSync('dist/theme.d.ts')) unlinkSync('dist/theme.d.ts');
+    if (existsSync('dist/theme.js.mjs')) unlinkSync('dist/theme.js.mjs');
+    if (existsSync('dist/theme.js.umd.js')) unlinkSync('dist/theme.js.umd.js');
+};
+
 // https://vitejs.dev/config/
 export default defineConfig({
     plugins: [
         vue(),
         dts({
             afterBuild: () => {
-                if (process.env.LIB_NAME === 'index') {
-                    copyFileSync('dist/index.d.ts', 'dist/index.d.mts');
-                }
-
-                if (process.env.LIB_NAME === 'styles') {
-                    setTimeout(function() {
-                        unlinkSync('dist/styles.d.ts');
-                        unlinkSync('dist/styles.js.mjs');
-                        unlinkSync('dist/styles.js.umd.js');
-                    }, 100);
-                }
+                setTimeout(removeUnnecessary, 50);
             },
         }),
     ],
@@ -67,17 +72,6 @@ export default defineConfig({
 
                     if (process.env.LIB_NAME === 'index') {
                         if (/\.css$/.test(assetInfo.name ?? '')) {
-
-                            if (fileName === 'index') {
-                                return `theme.css`;
-                            }
-
-                            if (/Theme$/.test(fileName)) {
-                                let folder = fileName.replace('Theme', '');
-
-                                return `components/${folder}/theme.css`;
-                            }
-
                             return `components/${fileName}/${fileName}.css`;
                         }
 
@@ -85,6 +79,10 @@ export default defineConfig({
                     }
 
                     if (process.env.LIB_NAME === 'styles') {
+                        return `styles/${fileName}.css`;
+                    }
+
+                    if (process.env.LIB_NAME === 'theme') {
                         return `styles/${fileName}.css`;
                     }
                 },
